Prevent excessive writes of .vscode/settings.json (fix issue #8)
The settings file currently is re-written twice every few seconds - for every folder in our workspace. This per se is excessive. But it's particularly bad if that folder is on a folder which is being synced to some cloud storage. It triggers sync activity for several files every few seconds. In one of my cases Dropbox basically was constantly busy syncing those files, as I have half a dozen or more folders in my workspace.

This change does still the same cleanup/patch routine before. But intermediate results are not written to and read from disc, but passed from one handler to the other in memory. Previously cleanup would remove the extension's entries from settings.json, write that change, and patch would add them back in if nothing has changed. But despite there being no change two writes were required.